+++ code/classeine-lib/clsn/ui/amatista/AmatistaUISkin.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "clsn/ui/UISkin.h"
+#include "clsn/ui/ui_skin.h"
 
 namespace clsn::ui::amatista
 {
-    class AmatistaUISkin : public UISkin
+    class AmatistaUISkin : public ui_skin
     {
     public:
         AmatistaUISkin();